create PhoneInput component, add comments
